# Conflicts:
#	test/main.cpp

+++ include/etl/flat_multiset.h
@@ -32,12 +32,12 @@ SOFTWARE.
 #define ETL_FLAT_MULTISET_INCLUDED
 
 #include "platform.h"
-#include "reference_flat_multiset.h"
-#include "pool.h"
-#include "placement_new.h"
-#include "nth_type.h"
-#include "type_traits.h"
 #include "initializer_list.h"
+#include "nth_type.h"
+#include "placement_new.h"
+#include "pool.h"
+#include "reference_flat_multiset.h"
+#include "type_traits.h"
 
 #include "private/comparator_is_transparent.h"
 
@@ -53,7 +53,8 @@ namespace etl
 {
   //***************************************************************************
   /// The base class for specifically sized flat_multisets.
-  /// Can be used as a reference type for all flat_multisets containing a specific type.
+  /// Can be used as a reference type for all flat_multisets containing a
+  /// specific type.
   ///\ingroup flat_multiset
   //***************************************************************************
   template <typename T, typename TKeyCompare = etl::less<T> >
@@ -62,8 +63,8 @@ namespace etl
   private:
 
     typedef etl::ireference_flat_multiset<T, TKeyCompare> refset_t;
-    typedef typename refset_t::lookup_t lookup_t;
-    typedef etl::ipool storage_t;
+    typedef typename refset_t::lookup_t                   lookup_t;
+    typedef etl::ipool                                    storage_t;
 
     typedef const T& key_parameter_t;
 
@@ -72,20 +73,20 @@ namespace etl
     typedef T                 key_type;
     typedef T                 value_type;
     typedef TKeyCompare       key_compare;
-    typedef value_type& reference;
+    typedef value_type&       reference;
     typedef const value_type& const_reference;
 #if ETL_USING_CPP11
     typedef value_type&& rvalue_reference;
 #endif
-    typedef value_type* pointer;
+    typedef value_type*       pointer;
     typedef const value_type* const_pointer;
     typedef size_t            size_type;
 
     typedef typename refset_t::iterator       iterator;
     typedef typename refset_t::const_iterator const_iterator;
 
-    typedef ETL_OR_STD::reverse_iterator<iterator>       reverse_iterator;
-    typedef ETL_OR_STD::reverse_iterator<const_iterator> const_reverse_iterator;
+    typedef ETL_OR_STD::reverse_iterator<iterator>                   reverse_iterator;
+    typedef ETL_OR_STD::reverse_iterator<const_iterator>             const_reverse_iterator;
     typedef typename etl::iterator_traits<iterator>::difference_type difference_type;
 
   public:
@@ -145,7 +146,8 @@ namespace etl
     }
 
     //*********************************************************************
-    /// Returns an reverse iterator to the reverse beginning of the flat_multiset.
+    /// Returns an reverse iterator to the reverse beginning of the
+    /// flat_multiset.
     ///\return Iterator to the reverse beginning of the flat_multiset.
     //*********************************************************************
     reverse_iterator rbegin()
@@ -154,7 +156,8 @@ namespace etl
     }
 
     //*********************************************************************
-    /// Returns a const reverse iterator to the reverse beginning of the flat_multiset.
+    /// Returns a const reverse iterator to the reverse beginning of the
+    /// flat_multiset.
     ///\return Const iterator to the reverse beginning of the flat_multiset.
     //*********************************************************************
     const_reverse_iterator rbegin() const
@@ -181,8 +184,10 @@ namespace etl
     }
 
     //*********************************************************************
-    /// Returns a const reverse iterator to the reverse beginning of the flat_multiset.
-    ///\return Const reverse iterator to the reverse beginning of the flat_multiset.
+    /// Returns a const reverse iterator to the reverse beginning of the
+    /// flat_multiset.
+    ///\return Const reverse iterator to the reverse beginning of the
+    /// flat_multiset.
     //*********************************************************************
     const_reverse_iterator crbegin() const
     {
@@ -200,8 +205,9 @@ namespace etl
 
     //*********************************************************************
     /// Assigns values to the flat_multiset.
-    /// If asserts or exceptions are enabled, emits flat_multiset_full if the flat_multiset does not have enough free space.
-    /// If asserts or exceptions are enabled, emits flat_multiset_iterator if the iterators are reversed.
+    /// If asserts or exceptions are enabled, emits flat_multiset_full if the
+    /// flat_multiset does not have enough free space. If asserts or exceptions
+    /// are enabled, emits flat_multiset_iterator if the iterators are reversed.
     ///\param first The iterator to the first element.
     ///\param last  The iterator to the last element + 1.
     //*********************************************************************
@@ -224,7 +230,8 @@ namespace etl
 
     //*********************************************************************
     /// Inserts a value to the flat_multiset.
-    /// If asserts or exceptions are enabled, emits flat_multiset_full if the flat_multiset is already full.
+    /// If asserts or exceptions are enabled, emits flat_multiset_full if the
+    /// flat_multiset is already full.
     ///\param value    The value to insert.
     //*********************************************************************
     ETL_OR_STD::pair<iterator, bool> insert(const_reference value)
@@ -238,7 +245,7 @@ namespace etl
       value_type* pvalue = storage.allocate<value_type>();
       ::new (pvalue) value_type(value);
       ETL_INCREMENT_DEBUG_COUNT;
-        result = refset_t::insert_at(i_element, *pvalue);
+      result = refset_t::insert_at(i_element, *pvalue);
 
       return result;
     }
@@ -246,7 +253,8 @@ namespace etl
 #if ETL_USING_CPP11
     //*********************************************************************
     /// Inserts a value to the flat_multiset.
-    /// If asserts or exceptions are enabled, emits flat_multiset_full if the flat_multiset is already full.
+    /// If asserts or exceptions are enabled, emits flat_multiset_full if the
+    /// flat_multiset is already full.
     ///\param value    The value to insert.
     //*********************************************************************
     ETL_OR_STD::pair<iterator, bool> insert(rvalue_reference value)
@@ -260,7 +268,7 @@ namespace etl
       value_type* pvalue = storage.allocate<value_type>();
       ::new (pvalue) value_type(etl::move(value));
       ETL_INCREMENT_DEBUG_COUNT;
-        result = refset_t::insert_at(i_element, *pvalue);
+      result = refset_t::insert_at(i_element, *pvalue);
 
       return result;
     }
@@ -268,7 +276,8 @@ namespace etl
 
     //*********************************************************************
     /// Inserts a value to the flat_multiset.
-    /// If asserts or exceptions are enabled, emits flat_multiset_full if the flat_multiset is already full.
+    /// If asserts or exceptions are enabled, emits flat_multiset_full if the
+    /// flat_multiset is already full.
     ///\param position The position to insert at.
     ///\param value    The value to insert.
     //*********************************************************************
@@ -280,7 +289,8 @@ namespace etl
 #if ETL_USING_CPP11
     //*********************************************************************
     /// Inserts a value to the flat_multiset.
-    /// If asserts or exceptions are enabled, emits flat_multiset_full if the flat_multiset is already full.
+    /// If asserts or exceptions are enabled, emits flat_multiset_full if the
+    /// flat_multiset is already full.
     ///\param position The position to insert at.
     ///\param value    The value to insert.
     //*********************************************************************
@@ -292,7 +302,8 @@ namespace etl
 
     //*********************************************************************
     /// Inserts a range of values to the flat_multiset.
-    /// If asserts or exceptions are enabled, emits flat_multiset_full if the flat_multiset does not have enough free space.
+    /// If asserts or exceptions are enabled, emits flat_multiset_full if the
+    /// flat_multiset does not have enough free space.
     ///\param position The position to insert at.
     ///\param first    The first element to add.
     ///\param last     The last + 1 element to add.
@@ -320,8 +331,8 @@ namespace etl
     /// Emplaces a value to the set.
     //*************************************************************************
 #if ETL_USING_CPP11 && ETL_NOT_USING_STLPORT && !defined(ETL_FLAT_MULTISET_FORCE_CPP03_IMPLEMENTATION)
-    template <typename ... Args>
-    ETL_OR_STD::pair<iterator, bool> emplace(Args && ... args)
+    template <typename... Args>
+    ETL_OR_STD::pair<iterator, bool> emplace(Args&&... args)
     {
       ETL_ASSERT(!full(), ETL_ERROR(flat_multiset_full));
 
@@ -440,7 +451,7 @@ namespace etl
       }
       else
       {
-        size_t d = etl::distance(range.first, range.second);
+        size_t d = static_cast<size_t>(etl::distance(range.first, range.second));
         erase(range.first, range.second);
         return d;
       }
@@ -459,7 +470,7 @@ namespace etl
       }
       else
       {
-        size_t d = etl::distance(range.first, range.second);
+        size_t d = static_cast<size_t>(etl::distance(range.first, range.second));
         erase(range.first, range.second);
         return d;
       }
@@ -492,8 +503,8 @@ namespace etl
 
     //*********************************************************************
     /// Erases a range of elements.
-    /// The range includes all the elements between first and last, including the
-    /// element pointed by first, but not the one pointed by last.
+    /// The range includes all the elements between first and last, including
+    /// the element pointed by first, but not the one pointed by last.
     ///\param first Iterator to the first element.
     ///\param last  Iterator to the last element.
     //*********************************************************************
@@ -517,7 +528,7 @@ namespace etl
     //*************************************************************************
     void clear()
     {
-      if ETL_IF_CONSTEXPR(etl::is_trivially_destructible<value_type>::value)
+      if ETL_IF_CONSTEXPR (etl::is_trivially_destructible<value_type>::value)
       {
         storage.release_all();
       }
@@ -534,7 +545,7 @@ namespace etl
       }
 
       ETL_RESET_DEBUG_COUNT;
-        refset_t::clear();
+      refset_t::clear();
     }
 
     //*********************************************************************
@@ -728,7 +739,7 @@ namespace etl
     //*************************************************************************
     /// Assignment operator.
     //*************************************************************************
-    iflat_multiset& operator = (const iflat_multiset& rhs)
+    iflat_multiset& operator=(const iflat_multiset& rhs)
     {
       if (&rhs != this)
       {
@@ -742,7 +753,7 @@ namespace etl
     //*************************************************************************
     /// Move assignment operator.
     //*************************************************************************
-    iflat_multiset& operator = (iflat_multiset&& rhs)
+    iflat_multiset& operator=(iflat_multiset&& rhs)
     {
       move_container(etl::move(rhs));
 
@@ -810,8 +821,8 @@ namespace etl
     /// Constructor.
     //*********************************************************************
     iflat_multiset(lookup_t& lookup_, storage_t& storage_)
-      : refset_t(lookup_),
-      storage(storage_)
+      : refset_t(lookup_)
+      , storage(storage_)
     {
     }
 
@@ -827,7 +838,7 @@ namespace etl
         this->clear();
 
         etl::iflat_multiset<T, TKeyCompare>::iterator first = rhs.begin();
-        etl::iflat_multiset<T, TKeyCompare>::iterator last = rhs.end();
+        etl::iflat_multiset<T, TKeyCompare>::iterator last  = rhs.end();
 
         // Move all of the elements.
         while (first != last)
@@ -854,19 +865,19 @@ namespace etl
     /// Internal debugging.
     ETL_DECLARE_DEBUG_COUNT;
 
-      //*************************************************************************
-      /// Destructor.
-      //*************************************************************************
+    //*************************************************************************
+    /// Destructor.
+    //*************************************************************************
 #if defined(ETL_POLYMORPHIC_FLAT_MULTISET) || defined(ETL_POLYMORPHIC_CONTAINERS)
+
   public:
-    virtual ~iflat_multiset()
-    {
-    }
+
+    virtual ~iflat_multiset() {}
 #else
+
   protected:
-    ~iflat_multiset()
-    {
-    }
+
+    ~iflat_multiset() {}
 #endif
   };
 
@@ -878,7 +889,7 @@ namespace etl
   ///\ingroup flat_multiset
   //***************************************************************************
   template <typename T, typename TKeyCompare>
-  bool operator ==(const etl::iflat_multiset<T, TKeyCompare>& lhs, const etl::iflat_multiset<T, TKeyCompare>& rhs)
+  bool operator==(const etl::iflat_multiset<T, TKeyCompare>& lhs, const etl::iflat_multiset<T, TKeyCompare>& rhs)
   {
     return (lhs.size() == rhs.size()) && etl::equal(lhs.begin(), lhs.end(), rhs.begin());
   }
@@ -891,7 +902,7 @@ namespace etl
   ///\ingroup flat_multiset
   //***************************************************************************
   template <typename T, typename TKeyCompare>
-  bool operator !=(const etl::iflat_multiset<T, TKeyCompare>& lhs, const etl::iflat_multiset<T, TKeyCompare>& rhs)
+  bool operator!=(const etl::iflat_multiset<T, TKeyCompare>& lhs, const etl::iflat_multiset<T, TKeyCompare>& rhs)
   {
     return !(lhs == rhs);
   }
@@ -976,7 +987,7 @@ namespace etl
     //*************************************************************************
     /// Assignment operator.
     //*************************************************************************
-    flat_multiset& operator = (const flat_multiset& rhs)
+    flat_multiset& operator=(const flat_multiset& rhs)
     {
       if (&rhs != this)
       {
@@ -990,7 +1001,7 @@ namespace etl
     //*************************************************************************
     /// Move assignment operator.
     //*************************************************************************
-    flat_multiset& operator = (flat_multiset&& rhs)
+    flat_multiset& operator=(flat_multiset&& rhs)
     {
       if (&rhs != this)
       {
@@ -1012,7 +1023,7 @@ namespace etl
     etl::vector<node_t*, MAX_SIZE> lookup;
   };
 
-  template <typename T,const size_t MAX_SIZE_, typename TCompare>
+  template <typename T, const size_t MAX_SIZE_, typename TCompare>
   ETL_CONSTANT size_t flat_multiset<T, MAX_SIZE_, TCompare>::MAX_SIZE;
 
   //*************************************************************************
@@ -1030,9 +1041,9 @@ namespace etl
   template <typename TKey, typename TKeyCompare = etl::less<TKey>, typename... T>
   constexpr auto make_flat_multiset(T&&... keys) -> etl::flat_multiset<TKey, sizeof...(T), TKeyCompare>
   {
-    return { etl::forward<T>(keys)... };
+    return {etl::forward<T>(keys)...};
   }
 #endif
-}
+} // namespace etl
 
 #endif
