+++ test/test_crc8_ebu.cpp
@@ -29,9 +29,9 @@ SOFTWARE.
 #include "unit_test_framework.h"
 
 #include <iterator>
+#include <stdint.h>
 #include <string>
 #include <vector>
-#include <stdint.h>
 
 #include "etl/crc8_ebu.h"
 #include "etl/endianness.h"
@@ -60,8 +60,8 @@ namespace
     //*************************************************************************
     TEST(test_crc8_ebu_4_constructor_constexpr)
     {
-      constexpr char data[] = "123456789";
-      constexpr uint8_t crc = etl::crc8_ebu_t4(data, data + 9);
+      constexpr char    data[] = "123456789";
+      constexpr uint8_t crc    = etl::crc8_ebu_t4(data, data + 9);
 
       CHECK_EQUAL(0x97U, int(crc));
     }
@@ -76,7 +76,7 @@ namespace
 
       for (size_t i = 0UL; i < data.size(); ++i)
       {
-        crc_calculator.add(data[i]);
+        crc_calculator.add(static_cast<uint8_t>(data[i]));
       }
 
       uint8_t crc = crc_calculator;
@@ -115,17 +115,17 @@ namespace
     //*************************************************************************
     TEST(test_crc8_ebu_4_add_range_endian)
     {
-      std::vector<uint8_t>  data1 = { 0x01U, 0x02U, 0x03U, 0x04U, 0x05U, 0x06U, 0x07U, 0x08U };
+      std::vector<uint8_t>  data1 = {0x01U, 0x02U, 0x03U, 0x04U, 0x05U, 0x06U, 0x07U, 0x08U};
       std::vector<uint32_t> data2;
       if (etl::endianness::value() == etl::endian::little)
       {
-        data2 = { 0x04030201UL, 0x08070605UL };
+        data2 = {0x04030201UL, 0x08070605UL};
       }
       else
       {
-        data2 = { 0x01020304UL, 0x05060708UL };
+        data2 = {0x01020304UL, 0x05060708UL};
       }
-      std::vector<uint8_t>  data3 = { 0x08U, 0x07U, 0x06U, 0x05U, 0x04U, 0x03U, 0x02U, 0x01U };
+      std::vector<uint8_t> data3 = {0x08U, 0x07U, 0x06U, 0x05U, 0x04U, 0x03U, 0x02U, 0x01U};
 
       uint8_t crc1 = etl::crc8_ebu_t4(data1.begin(), data1.end());
       uint8_t crc2 = etl::crc8_ebu_t4((uint8_t*)&data2[0], (uint8_t*)(&data2[0] + data2.size()));
@@ -151,8 +151,8 @@ namespace
     //*************************************************************************
     TEST(test_crc8_ebu_16_constructor_constexpr)
     {
-      constexpr char data[] = "123456789";
-      constexpr uint8_t crc = etl::crc8_ebu_t16(data, data + 9);
+      constexpr char    data[] = "123456789";
+      constexpr uint8_t crc    = etl::crc8_ebu_t16(data, data + 9);
 
       CHECK_EQUAL(0x97U, int(crc));
     }
@@ -167,7 +167,7 @@ namespace
 
       for (size_t i = 0UL; i < data.size(); ++i)
       {
-        crc_calculator.add(data[i]);
+        crc_calculator.add(static_cast<uint8_t>(data[i]));
       }
 
       uint8_t crc = crc_calculator;
@@ -206,17 +206,17 @@ namespace
     //*************************************************************************
     TEST(test_crc8_ebu_16_add_range_endian)
     {
-      std::vector<uint8_t>  data1 = { 0x01U, 0x02U, 0x03U, 0x04U, 0x05U, 0x06U, 0x07U, 0x08U };
+      std::vector<uint8_t>  data1 = {0x01U, 0x02U, 0x03U, 0x04U, 0x05U, 0x06U, 0x07U, 0x08U};
       std::vector<uint32_t> data2;
       if (etl::endianness::value() == etl::endian::little)
       {
-        data2 = { 0x04030201UL, 0x08070605UL };
+        data2 = {0x04030201UL, 0x08070605UL};
       }
       else
       {
-        data2 = { 0x01020304UL, 0x05060708UL };
+        data2 = {0x01020304UL, 0x05060708UL};
       }
-      std::vector<uint8_t>  data3 = { 0x08U, 0x07U, 0x06U, 0x05U, 0x04U, 0x03U, 0x02U, 0x01U };
+      std::vector<uint8_t> data3 = {0x08U, 0x07U, 0x06U, 0x05U, 0x04U, 0x03U, 0x02U, 0x01U};
 
       uint8_t crc1 = etl::crc8_ebu_t16(data1.begin(), data1.end());
       uint8_t crc2 = etl::crc8_ebu_t16((uint8_t*)&data2[0], (uint8_t*)(&data2[0] + data2.size()));
@@ -242,8 +242,8 @@ namespace
     //*************************************************************************
     TEST(test_crc8_ebu_256_constructor_constexpr)
     {
-      constexpr char data[] = "123456789";
-      constexpr uint8_t crc = etl::crc8_ebu(data, data + 9);
+      constexpr char    data[] = "123456789";
+      constexpr uint8_t crc    = etl::crc8_ebu(data, data + 9);
 
       CHECK_EQUAL(0x97U, int(crc));
     }
@@ -258,7 +258,7 @@ namespace
 
       for (size_t i = 0UL; i < data.size(); ++i)
       {
-        crc_calculator.add(data[i]);
+        crc_calculator.add(static_cast<uint8_t>(data[i]));
       }
 
       uint8_t crc = crc_calculator;
@@ -297,17 +297,17 @@ namespace
     //*************************************************************************
     TEST(test_crc8_ebu_256_add_range_endian)
     {
-      std::vector<uint8_t>  data1 = { 0x01U, 0x02U, 0x03U, 0x04U, 0x05U, 0x06U, 0x07U, 0x08U };
+      std::vector<uint8_t>  data1 = {0x01U, 0x02U, 0x03U, 0x04U, 0x05U, 0x06U, 0x07U, 0x08U};
       std::vector<uint32_t> data2;
       if (etl::endianness::value() == etl::endian::little)
       {
-        data2 = { 0x04030201UL, 0x08070605UL };
+        data2 = {0x04030201UL, 0x08070605UL};
       }
       else
       {
-        data2 = { 0x01020304UL, 0x05060708UL };
+        data2 = {0x01020304UL, 0x05060708UL};
       }
-      std::vector<uint8_t>  data3 = { 0x08U, 0x07U, 0x06U, 0x05U, 0x04U, 0x03U, 0x02U, 0x01U };
+      std::vector<uint8_t> data3 = {0x08U, 0x07U, 0x06U, 0x05U, 0x04U, 0x03U, 0x02U, 0x01U};
 
       uint8_t crc1 = etl::crc8_ebu(data1.begin(), data1.end());
       uint8_t crc2 = etl::crc8_ebu((uint8_t*)&data2[0], (uint8_t*)(&data2[0] + data2.size()));
@@ -317,5 +317,4 @@ namespace
       CHECK_EQUAL(int(crc1), int(crc3));
     }
   }
-}
-
+} // namespace
