+++ test/test_unordered_multiset.cpp
@@ -28,39 +28,39 @@ SOFTWARE.
 
 #include "unit_test_framework.h"
 
-#include <set>
-#include <array>
 #include <algorithm>
-#include <utility>
+#include <array>
 #include <iterator>
-#include <string>
-#include <vector>
 #include <numeric>
+#include <set>
+#include <string>
 #include <unordered_set>
+#include <utility>
+#include <vector>
 
 #include "data.h"
 
-#include "etl/unordered_multiset.h"
 #include "etl/checksum.h"
+#include "etl/unordered_multiset.h"
 
 namespace
 {
   using DC  = TestDataDC<std::string>;
   using NDC = TestDataNDC<std::string>;
-}
+} // namespace
 
 namespace etl
 {
   template <>
   struct hash<NDC>
   {
-    size_t operator ()(const NDC& e) const
+    size_t operator()(const NDC& e) const
     {
       size_t sum = 0U;
       return std::accumulate(e.value.begin(), e.value.end(), sum);
     }
   };
-}
+} // namespace etl
 
 namespace
 {
@@ -77,7 +77,7 @@ namespace
     {
     }
 
-    size_t operator ()(uint32_t e) const
+    size_t operator()(uint32_t e) const
     {
       return size_t(e);
     }
@@ -98,7 +98,7 @@ namespace
     {
     }
 
-    size_t operator ()(uint32_t lhs, uint32_t rhs) const
+    size_t operator()(uint32_t lhs, uint32_t rhs) const
     {
       return (lhs == rhs);
     }
@@ -112,7 +112,10 @@ namespace
   {
     size_t modulus;
 
-    parameterized_hash(size_t modulus_ = 2) : modulus(modulus_){}
+    parameterized_hash(size_t modulus_ = 2)
+      : modulus(modulus_)
+    {
+    }
 
     size_t operator()(size_t val) const
     {
@@ -127,7 +130,10 @@ namespace
     size_t modulus;
 
     // Hasher whose hash behaviour depends on provided data.
-    parameterized_equal(size_t modulus_ = 2) : modulus(modulus_){}
+    parameterized_equal(size_t modulus_ = 2)
+      : modulus(modulus_)
+    {
+    }
 
     bool operator()(size_t lhs, size_t rhs) const
     {
@@ -140,14 +146,14 @@ namespace
   {
     typedef int is_transparent;
 
-    size_t operator ()(const char* s) const
+    size_t operator()(const char* s) const
     {
-      return std::accumulate(s, s + etl::strlen(s), 0);
+      return std::accumulate(s, s + etl::strlen(s), size_t(0));
     }
 
-    size_t operator ()(const std::string& s) const
+    size_t operator()(const std::string& s) const
     {
-      return std::accumulate(s.begin(), s.end(), 0);
+      return std::accumulate(s.begin(), s.end(), size_t(0));
     }
   };
 
@@ -160,12 +166,12 @@ namespace
 
     struct simple_hash
     {
-      size_t operator ()(const NDC& value) const
+      size_t operator()(const NDC& value) const
       {
         return etl::checksum<size_t>(value.value.begin(), value.value.end());
       }
 
-      size_t operator ()(const ItemM& value) const
+      size_t operator()(const ItemM& value) const
       {
         etl::checksum<size_t> sum;
 
@@ -178,7 +184,7 @@ namespace
 
     using DataM = etl::unordered_multiset<ItemM, SIZE, SIZE, simple_hash>;
 
-    using DataDC          = etl::unordered_multiset<DC,  SIZE, SIZE / 2, simple_hash>;
+    using DataDC          = etl::unordered_multiset<DC, SIZE, SIZE / 2, simple_hash>;
     using DataNDC         = etl::unordered_multiset<NDC, SIZE, SIZE / 2, simple_hash>;
     using IDataNDC        = etl::iunordered_multiset<NDC, simple_hash>;
     using DataTransparent = etl::unordered_multiset<std::string, SIZE, SIZE / 2, transparent_hash, etl::equal_to<>>;
@@ -235,25 +241,13 @@ namespace
     {
       SetupFixture()
       {
-        NDC n[] =
-        {
-          N0, N1, N2, N3, N4, N5, N6, N7, N8, N9
-        };
+        NDC n[] = {N0, N1, N2, N3, N4, N5, N6, N7, N8, N9};
 
-        NDC n2[] =
-        {
-          N0, N1, N2, N3, N4, N5, N6, N7, N8, N9, N10
-        };
+        NDC n2[] = {N0, N1, N2, N3, N4, N5, N6, N7, N8, N9, N10};
 
-        NDC n3[] =
-        {
-          N10, N11, N12, N13, N14, N15, N16, N17, N18, N19
-        };
+        NDC n3[] = {N10, N11, N12, N13, N14, N15, N16, N17, N18, N19};
 
-        NDC n4[] =
-        {
-          N0, N1, N1, N1, N2
-        };
+        NDC n4[] = {N0, N1, N1, N1, N2};
 
         initial_data.assign(std::begin(n), std::end(n));
         excess_data.assign(std::begin(n2), std::end(n2));
@@ -277,8 +271,8 @@ namespace
     //*************************************************************************
     TEST_FIXTURE(SetupFixture, test_cpp17_deduced_constructor)
     {
-      etl::unordered_multiset data{ N0, N1, N2, N3, N4, N5, N6, N7, N8, N9 };
-      etl::unordered_multiset<NDC, 10U> check = { N0, N1, N2, N3, N4, N5, N6, N7, N8, N9 };
+      etl::unordered_multiset           data{N0, N1, N2, N3, N4, N5, N6, N7, N8, N9};
+      etl::unordered_multiset<NDC, 10U> check = {N0, N1, N2, N3, N4, N5, N6, N7, N8, N9};
 
       CHECK(!data.empty());
       CHECK(data.full());
@@ -436,7 +430,7 @@ namespace
     //*************************************************************************
     TEST_FIXTURE(SetupFixture, test_assign_range_using_transparent_comparator)
     {
-      std::array<const char*, 8> initial = { "AA", "BB", "CC", "DD", "EE", "FF", "GG", "HH" };
+      std::array<const char*, 8> initial = {"AA", "BB", "CC", "DD", "EE", "FF", "GG", "HH"};
 
       DataTransparent data;
 
@@ -448,7 +442,7 @@ namespace
       {
         idata = data.find(initial[i]);
         CHECK(idata != data.end());
-      } 
+      }
     }
 
     //*************************************************************************
@@ -545,7 +539,7 @@ namespace
     //*************************************************************************
     TEST_FIXTURE(SetupFixture, test_insert_range_using_transparent_comparator)
     {
-      std::array<const char*, 8> initial = { "AA", "BB", "CC", "DD", "EE", "FF", "GG", "HH" };
+      std::array<const char*, 8> initial = {"AA", "BB", "CC", "DD", "EE", "FF", "GG", "HH"};
 
       DataTransparent data;
 
@@ -557,7 +551,7 @@ namespace
       {
         idata = data.find(initial[i]);
         CHECK(idata != data.end());
-      } 
+      }
     }
 
     //*************************************************************************
@@ -615,7 +609,7 @@ namespace
     //*************************************************************************
     TEST_FIXTURE(SetupFixture, test_erase_key_using_transparent_comparator)
     {
-      std::array<std::string, 8> initial = { "AA", "BB", "CC", "DD", "EE", "FF", "GG", "HH" };
+      std::array<std::string, 8> initial = {"AA", "BB", "CC", "DD", "EE", "FF", "GG", "HH"};
 
       DataTransparent data(initial.begin(), initial.end());
 
@@ -637,7 +631,7 @@ namespace
       ++inext;
 
       DataNDC::iterator iafter = data.erase(idata);
-      idata = data.find(N5);
+      idata                    = data.find(N5);
 
       CHECK(idata == data.end());
       CHECK(inext == iafter);
@@ -657,7 +651,7 @@ namespace
       ++inext;
 
       DataNDC::const_iterator iafter = data.erase(idata);
-      idata = data.find(N5);
+      idata                          = data.find(N5);
 
       CHECK(idata == data.cend());
       CHECK(inext == iafter);
@@ -790,7 +784,7 @@ namespace
     //*************************************************************************
     TEST_FIXTURE(SetupFixture, test_count_key_using_transparent_comparator)
     {
-      std::array<std::string, 8> initial = { "AA", "BB", "CC", "DD", "EE", "FF", "GG", "HH" };
+      std::array<std::string, 8> initial = {"AA", "BB", "CC", "DD", "EE", "FF", "GG", "HH"};
 
       DataTransparent data(initial.begin(), initial.end());
 
@@ -892,8 +886,8 @@ namespace
     //*************************************************************************
     TEST_FIXTURE(SetupFixture, test_hash_function)
     {
-      DataNDC data;
-      DataNDC::hasher hash_function =  data.hash_function();
+      DataNDC         data;
+      DataNDC::hasher hash_function = data.hash_function();
 
       CHECK_EQUAL(simple_hash()(NDC(std::string("ABCDEF"))), hash_function(NDC(std::string("ABCDEF"))));
     }
@@ -901,7 +895,7 @@ namespace
     //*************************************************************************
     TEST_FIXTURE(SetupFixture, test_key_eq_function)
     {
-      DataNDC data;
+      DataNDC            data;
       DataNDC::key_equal key_eq = data.key_eq();
 
       CHECK(key_eq(NDC(std::string("ABCDEF")), NDC(std::string("ABCDEF"))));
@@ -916,7 +910,7 @@ namespace
       CHECK_CLOSE(0.0, data.load_factor(), 0.01);
 
       // Half the buckets used.
-      data.assign(initial_data.begin(), initial_data.begin() + (initial_data.size() / 4));
+      data.assign(initial_data.begin(), initial_data.begin() + static_cast<ptrdiff_t>(initial_data.size() / 4));
       CHECK_CLOSE(0.4, data.load_factor(), 0.01);
 
       // All of the buckets used.
@@ -933,12 +927,12 @@ namespace
         // Force hash collisions
         size_t operator()(int key) const
         {
-          return key % 4;
+          return static_cast<size_t>(key % 4);
         }
       };
 
-      std::vector<int> random_keys1 = { 17, 14, 3,  7, 2, 6, 9,  3, 18, 10,  8, 11,  4, 1, 12, 15, 16,  0,  5, 19 };
-      std::vector<int> random_keys2 = { 3,  6, 5, 17, 2, 7, 3, 19,  8, 15, 14,  0, 18, 4, 10,  9, 16, 11, 12,  1 };
+      std::vector<int> random_keys1 = {17, 14, 3, 7, 2, 6, 9, 3, 18, 10, 8, 11, 4, 1, 12, 15, 16, 0, 5, 19};
+      std::vector<int> random_keys2 = {3, 6, 5, 17, 2, 7, 3, 19, 8, 15, 14, 0, 18, 4, 10, 9, 16, 11, 12, 1};
 
       // Check that the input data is valid.
       CHECK_EQUAL(random_keys1.size(), random_keys2.size());
@@ -1015,7 +1009,7 @@ namespace
       CustomHashFunction chf1(1);
       CustomKeyEq        ceq2(2);
 
-      std::array<uint32_t, 5> data = { 1, 2, 3, 4, 5 };
+      std::array<uint32_t, 5> data = {1, 2, 3, 4, 5};
 
       etl::unordered_multiset<uint32_t, 5, 5, CustomHashFunction, CustomKeyEq> set1(data.begin(), data.end(), chf1, ceq2);
 
@@ -1029,7 +1023,7 @@ namespace
       CustomHashFunction chf1(1);
       CustomKeyEq        ceq2(2);
 
-      etl::unordered_multiset<uint32_t, 5, 5, CustomHashFunction, CustomKeyEq> set1({ 1, 2, 3, 4, 5 }, chf1, ceq2);
+      etl::unordered_multiset<uint32_t, 5, 5, CustomHashFunction, CustomKeyEq> set1({1, 2, 3, 4, 5}, chf1, ceq2);
 
       CHECK_EQUAL(chf1.id, set1.hash_function().id);
       CHECK_EQUAL(ceq2.id, set1.key_eq().id);
@@ -1045,8 +1039,8 @@ namespace
     TEST(test_parameterized_eq)
     {
       constexpr std::size_t MODULO = 4;
-      parameterized_hash hash{MODULO};
-      parameterized_equal eq{MODULO};
+      parameterized_hash    hash{MODULO};
+      parameterized_equal   eq{MODULO};
       // values are equal modulo 4
       etl::unordered_multiset<std::size_t, 10, 10, parameterized_hash, parameterized_equal> set;
       set.insert(2);
@@ -1085,7 +1079,7 @@ namespace
     {
       DataNDC data(initial_data.begin(), initial_data.end());
 
-      NDC not_inserted  = NDC("ZZ");
+      NDC not_inserted = NDC("ZZ");
 
       CHECK_TRUE(data.contains(N0));
       CHECK_FALSE(data.contains(not_inserted));
@@ -1094,14 +1088,14 @@ namespace
     //*************************************************************************
     TEST(test_contains_with_transparent_comparator)
     {
-      std::array<const char*, 8> initial = { "AA", "BB", "CC", "DD", "EE", "FF", "GG", "HH" };
+      std::array<const char*, 8> initial = {"AA", "BB", "CC", "DD", "EE", "FF", "GG", "HH"};
 
       DataTransparent data(initial.begin(), initial.end());
 
-      const char* not_inserted  = "ZZ";
+      const char* not_inserted = "ZZ";
 
       CHECK_TRUE(data.contains("FF"));
       CHECK_FALSE(data.contains(not_inserted));
     }
   }
-}
+} // namespace
